+++ compose/ui/ui/src/uikitMain/kotlin/androidx/compose/ui/platform/UIKitTextInputService.uikit.kt
@@ -22,8 +22,8 @@ import androidx.compose.ui.input.key.NativeKeyEvent
 import androidx.compose.ui.text.input.*
 import androidx.compose.ui.unit.Density
 import androidx.compose.ui.window.IntermediateTextInputUIView
-import androidx.compose.ui.window.di.FocusStack
-import androidx.compose.ui.window.di.KeyboardEventHandler
+import androidx.compose.ui.window.FocusStack
+import androidx.compose.ui.window.KeyboardEventHandler
 import kotlin.math.absoluteValue
 import kotlin.math.min
 import org.jetbrains.skia.BreakIterator
